WIP - Start moving Clap code to a crate for clap-markdown to be run as an xtask.
