+++ crates/weaver_cli/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+name = "weaver_cli"
+version.workspace = true
+authors.workspace = true
+repository.workspace = true
+license.workspace = true
+publish.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+
+[dependencies]
+clap = { version = "4.5.37", features = ["derive"] }
+clap_complete = "4.5.50"
+#serde_yaml = "0.9.34"
+weaver_common = { path = "../weaver_common" }
+weaver_emit = { path = "../weaver_emit" }
+serde_yaml = "0.9.34+deprecated"
+thiserror = "1.0.69"
+serde = { version = "1.0.219", features = ["derive"] }
+miette = { version = "7.5.0", features = ["fancy", "serde"] }
+
+[lints]
+workspace = true
