+++ crates/weaver_cli/README.md
@@ -1,0 +1,4 @@
+# Weaver CLI
+
+This crate is used to isolate the CLI code from the rest of the weaver codebase. This is so we can generate the markdown
+help, as well as the man page file without having to build that into the weaver binary.
