+++ crates/xtask/Cargo.toml
@@ -20,6 +20,8 @@ gix = { version = "0.71.0", default-features = false, features = [
 ] }
 semver = "1.0.26"
 toml = "0.8.19"
+clap-markdown = "0.1.5"
+weaver_cli = { path = "../weaver_cli" }
 
 #[lints]
 #workspace = true
